re-struction patient list component - create new directory and move subcomponents
