+++ patientsearch/src/__tests__/components/patientList/DetailPanel.test.js
@@ -1,10 +1,15 @@
 import { shallow } from "enzyme";
 import React from "react";
-import DetailPanel from "../../js/components/DetailPanel";
+import DetailPanel from "../../../js/components/patientList/DetailPanel";
 
 describe("DetailPanel", () => {
   it("DetailPanel component renders without crashing", () => {
-    const wrapper = shallow(<DetailPanel />);
+    const data = {
+      rowData: {
+        id: 1
+      }
+    }
+    const wrapper = shallow(<DetailPanel data={data}/>);
     expect(wrapper).toBeDefined();
   });
 });
