+++ patientsearch/src/js/components/patientList/Legend.js
@@ -1,0 +1,14 @@
+import useStyles from "../../../styles/patientListStyle";
+import { usePatientListContext } from "../../context/PatientListContextProvider";
+
+export default function Legend() {
+  const classes = useStyles();
+  const { containNoPMPRow } = usePatientListContext();
+  if (containNoPMPRow)
+    return (
+      <div className={classes.legend}>
+        <span className={classes.legendIcon}></span> Not in PMP
+      </div>
+    );
+  return <div className={classes.spacer}></div>;
+}
