+++ patientsearch/src/js/components/patientList/MyPatientsCheckbox.js
@@ -3,9 +3,13 @@ import { makeStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
-import ErrorIcon from '@material-ui/icons/ReportProblemOutlined';
+import ErrorIcon from "@material-ui/icons/ReportProblemOutlined";
 import Tooltip from "@material-ui/core/Tooltip";
 import Typography from "@material-ui/core/Typography";
+import {usePatientListContext} from "../../context/PatientListContextProvider";
+import {
+  getPatientIdsByCareTeamParticipant,
+} from "../../helpers/utility";
 
 const checkBoxStyles = makeStyles((theme) => {
   return {
@@ -15,11 +19,11 @@ const checkBoxStyles = makeStyles((theme) => {
     warningBg: {
       backgroundColor: theme.palette.warning.dark,
       color: "#FFF",
-      fontSize: "0.95rem"
+      fontSize: "0.95rem",
     },
     warning: {
-      color: theme.palette.warning.dark
-    }
+      color: theme.palette.warning.dark,
+    },
   };
 });
 const formControlStyles = makeStyles((theme) => {
@@ -27,19 +31,48 @@ const formControlStyles = makeStyles((theme) => {
     root: {
       backgroundColor: "#f7f7f7",
       marginLeft: theme.spacing(1),
-      marginRight: theme.spacing(1)
+      marginRight: theme.spacing(1),
     },
   };
 });
 
-export default function MyPatientsCheckbox({ label, shouldDisable, shouldCheck, changeEvent, error }) {
+export default function MyPatientsCheckbox({
+  shouldDisable,
+  changeEvent,
+}) {
+  const {
+    getAppSettingByKey,
+    patientIdsByCareTeamParticipant,
+    setPatientIdsByCareTeamParticipant,
+    user,
+    userError,
+    tableRef
+  } = usePatientListContext();
   const checkboxClasses = checkBoxStyles();
   const formControlClasses = formControlStyles();
-  const [state, setState] = useState(shouldCheck);
+  const [state, setState] = useState(
+    patientIdsByCareTeamParticipant && patientIdsByCareTeamParticipant.length
+  );
   const handleChange = (event) => {
     setState(event.target.checked);
-    if (changeEvent) changeEvent(event.target.checked);
+    if (!event.target.checked) {
+      setPatientIdsByCareTeamParticipant(null);
+      if (tableRef.current) tableRef.current.onQueryChange();
+      if (changeEvent) changeEvent();
+      return;
+    }
+    // NOTE - retrieving id(s) of patients whose care team the practitioner is part of
+    getPatientIdsByCareTeamParticipant(user ? user.practitionerId : null).then(
+      (result) => {
+        setPatientIdsByCareTeamParticipant(
+          result && result.length ? result : [-1]
+        );
+        if (tableRef.current) tableRef.current.onQueryChange();
+        if (changeEvent) changeEvent();
+      }
+    );
   };
+  if (!getAppSettingByKey("ENABLE_PROVIDER_FILTER")) return false;
   return (
     <div style={{ display: "flex", alignItems: "center", gap: "8px" }}>
       <FormControlLabel
@@ -59,19 +92,25 @@ export default function MyPatientsCheckbox({ label, shouldDisable, shouldCheck, 
             }}
           />
         }
-        label={<Typography variant="body2">{label}</Typography>}
+        label={
+          <Typography variant="body2">
+            {getAppSettingByKey("MY_PATIENTS_FILTER_LABEL")}
+          </Typography>
+        }
       />
-      {error && (
+      {userError && (
         <Tooltip
-          title={error}
+          title={userError}
           enterTouchDelay={0}
           classes={{
             tooltip: checkboxClasses.warningBg,
           }}
         >
-          <ErrorIcon classes={{
-            root: checkboxClasses.warning
-          }} />
+          <ErrorIcon
+            classes={{
+              root: checkboxClasses.warning,
+            }}
+          />
         </Tooltip>
       )}
     </div>
@@ -79,9 +118,6 @@ export default function MyPatientsCheckbox({ label, shouldDisable, shouldCheck, 
 }
 
 MyPatientsCheckbox.propTypes = {
-  label: PropTypes.string.isRequired,
-  shouldCheck: PropTypes.bool,
   shouldDisable: PropTypes.bool,
   changeEvent: PropTypes.func,
-  error: PropTypes.string
 };
