+++ patientsearch/src/js/components/patientList/main.js
@@ -1,0 +1,10 @@
+import PatientListContextProvider from "../../context/PatientListContextProvider";
+import PatientListTable from "./PatientListTable";
+
+export default function MainContent() {
+  return (
+    <PatientListContextProvider>
+      <PatientListTable></PatientListTable>
+    </PatientListContextProvider>
+  );
+}
